Refactor endpoint related structs to own packages
