+++ handler/generator/response.go
@@ -1,4 +1,4 @@
-package handler
+package generator
 
 import (
 	"io"
@@ -6,16 +6,6 @@ import (
 
 	"github.com/hashicorp/hcl/v2"
 )
-
-type Request struct {
-	Backend *Backend
-	Body    io.Reader
-	Context hcl.Body
-	// Dispatch bool
-	Method string
-	Name   string // label
-	URL    string
-}
 
 type Response struct {
 	Body      io.Reader
