+++ handler/transport/backend.go
@@ -1,4 +1,4 @@
-package handler
+package transport
 
 import (
 	"net/http"
@@ -11,14 +11,18 @@ var _ http.RoundTripper = &Backend{}
 type Backend struct {
 	context       hcl.Body
 	name          string
-	transportConf *TransportConfig
+	transportConf *Config
 	AccessControl string // maps to basic-auth atm
-	OpenAPI       *OpenAPIValidatorOptions
+	//OpenAPI       *OpenAPIValidatorOptions
 	// oauth
 	// ...
 	// TODO: OrderedList for origin AC, middlewares etc.
 }
 
+func NewBackend(conf *Config) *Backend {
+	return &Backend{transportConf: conf}
+}
+
 func (b *Backend) RoundTrip(req *http.Request) (*http.Response, error) {
-	return getTransport(b.transportConf).RoundTrip(req)
+	return Get(b.transportConf).RoundTrip(req)
 }
